+++ pkg/helm/controller/reconcile.go
@@ -16,14 +16,17 @@ package controller
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"time"
 
 	rpb "helm.sh/helm/v3/pkg/release"
+	"helm.sh/helm/v3/pkg/storage/driver"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -126,7 +129,7 @@ func (r HelmOperatorReconciler) Reconcile(request reconcile.Request) (reconcile.
 		}
 
 		uninstalledRelease, err := manager.UninstallRelease(context.TODO())
-		if err != nil && err != release.ErrNotFound {
+		if err != nil && !errors.Is(err, driver.ErrReleaseNotFound) {
 			log.Error(err, "Failed to uninstall release")
 			status.SetCondition(types.HelmAppCondition{
 				Type:    types.ConditionReleaseFailed,
@@ -139,7 +142,7 @@ func (r HelmOperatorReconciler) Reconcile(request reconcile.Request) (reconcile.
 		}
 		status.RemoveCondition(types.ConditionReleaseFailed)
 
-		if err == release.ErrNotFound {
+		if errors.Is(err, driver.ErrReleaseNotFound) {
 			log.Info("Release not found, removing finalizer")
 		} else {
 			log.Info("Uninstalled release")
@@ -154,6 +157,7 @@ func (r HelmOperatorReconciler) Reconcile(request reconcile.Request) (reconcile.
 			status.DeployedRelease = nil
 		}
 		if err := r.updateResourceStatus(o, status); err != nil {
+			log.Info("Failed to update CR status")
 			return reconcile.Result{}, err
 		}
 
@@ -164,10 +168,21 @@ func (r HelmOperatorReconciler) Reconcile(request reconcile.Request) (reconcile.
 			}
 		}
 		o.SetFinalizers(finalizers)
-		err = r.updateResource(o)
+		if err := r.updateResource(o); err != nil {
+			log.Info("Failed to remove CR uninstall finalizer")
+			return reconcile.Result{}, err
+		}
 
-		// Need to requeue because finalizer update does not change metadata.generation
-		return reconcile.Result{Requeue: true}, err
+		// Since the client is hitting a cache, waiting for the
+		// deletion here will guarantee that the next reconciliation
+		// will see that the CR has been deleted and that there's
+		// nothing left to do.
+		if err := r.waitForDeletion(o); err != nil {
+			log.Info("Failed waiting for CR deletion")
+			return reconcile.Result{}, err
+		}
+
+		return reconcile.Result{}, nil
 	}
 
 	if !manager.IsInstalled() {
@@ -311,6 +326,26 @@ func (r HelmOperatorReconciler) updateResource(o runtime.Object) error {
 func (r HelmOperatorReconciler) updateResourceStatus(o *unstructured.Unstructured, status *types.HelmAppStatus) error {
 	o.Object["status"] = status
 	return r.Client.Status().Update(context.TODO(), o)
+}
+
+func (r HelmOperatorReconciler) waitForDeletion(o runtime.Object) error {
+	key, err := client.ObjectKeyFromObject(o)
+	if err != nil {
+		return err
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
+	defer cancel()
+	return wait.PollImmediateUntil(time.Millisecond*10, func() (bool, error) {
+		err := r.Client.Get(ctx, key, o)
+		if apierrors.IsNotFound(err) {
+			return true, nil
+		}
+		if err != nil {
+			return false, err
+		}
+		return false, nil
+	}, ctx.Done())
 }
 
 func contains(l []string, s string) bool {